Fiwix64: make __uid_t/__gid_t 32-bit (musl x86-64 ABI) + wire getgroups/setgroups
The kernel's __uid_t/__gid_t were __u16 and struct proc's uid/gid/euid/
egid/suid/sgid were unsigned short, but musl x86-64 uid_t/gid_t are
32-bit. Widening aligns every uid/gid syscall signature and struct ipc_perm
with the ABI (the ipc64 conversion becomes a 32->32 identity instead of
silently narrowing).

- types.h: __uid_t/__gid_t -> __u32
- process.h: struct proc uid/gid/euid/egid/suid/sgid -> __u32
- process.c proc_slot_init: initialize groups[] to -1 (was uninitialized
  garbage from the memset-to-0 proc table, which the -1 sentinel could
  never match - a group-permission bypass in check_group)
- fork.c: restore parent's groups after proc_slot_init (which wipes them)
- syscalls.c: wire getgroups(115)/setgroups(116) - now ABI-correct
- getgroups.c: count ngroups first; EINVAL on size<0 or size<ngroups;
  verify exactly ngroups*sizeof(__gid_t) (no 32-bit size*4 wrap); the
  n>size off-by-one is fixed (n>=size)
- setgroups.c: EINVAL/EPERM checks before check_user_area (bounded);
  skip verify when size==0 so setgroups(0,NULL) can clear the list

Verified: getgroups/setgroups round-trip (1000,1001), fork inheritance
(child sees parent's groups), getgroups(0)=count, getgroups(1)/getgroups(-1)
return EINVAL, setgroups(0,NULL) clears; kill9 LOOP_DONE; test_mmap 22/22;
full stress ALL DONE 0 failures.

+++ include/fiwix/process.h
@@ -129,12 +129,12 @@ struct proc {
 	__time_t start_time;
 	int exit_code;	
 	void *sleep_address;
-	unsigned short int uid;		/* real user ID */
-	unsigned short int gid;		/* real group ID */
-	unsigned short int euid;	/* effective user ID */
-	unsigned short int egid;	/* effective group ID */
-	unsigned short int suid;	/* saved user ID */
-	unsigned short int sgid;	/* saved group ID */
+	__u32 uid;			/* real user ID */
+	__u32 gid;			/* real group ID */
+	__u32 euid;			/* effective user ID */
+	__u32 egid;			/* effective group ID */
+	__u32 suid;			/* saved user ID */
+	__u32 sgid;			/* saved group ID */
 	unsigned short int fd[OPEN_MAX];
 	unsigned char fd_flags[OPEN_MAX];
 	struct inode *root;
